+++ packages/extras/CHANGELOG.md
@@ -1,5 +1,51 @@
 # @threlte/extras
 
+## 7.5.0
+
+### Minor Changes
+
+- bcc96f23: Added `<AnimatedSpriteMaterial>`
+
+## 7.4.0
+
+### Minor Changes
+
+- 20e8711a: Added `<CSM>` to extras
+
+## 7.3.0
+
+### Minor Changes
+
+- c2ddef77: Added `suspended` store to return type of `useSuspense`
+
+## 7.2.0
+
+### Minor Changes
+
+- 206fbc83: Added suspense support to `<GLTF>` and `<Environment>`
+
+### Patch Changes
+
+- 79a7fa4f: Aligned `onReveal` with `onMount` in that you may return a cleanup function that runs whenever a component is resuspended or unmounted
+
+## 7.1.1
+
+### Patch Changes
+
+- 060fbbdf: Changed eventDispatcher to rawEventDispatcher in `<Suspense>`, added type def
+
+## 7.1.0
+
+### Minor Changes
+
+- cd27881c: Added suspend event to `<Suspense>`
+
+## 7.0.1
+
+### Patch Changes
+
+- 62d3f395: Add ref binding to `<InstancedMesh>`
+
 ## 7.0.0
 
 ### Major Changes