+++ packages/xr/CHANGELOG.md
@@ -1,5 +1,17 @@
 # @threlte/xr
 
+## 0.0.11
+
+### Patch Changes
+
+- 8ebaa97b: Force svelte/motion to use session.requestAnimationFrame when presenting
+
+## 0.0.10
+
+### Patch Changes
+
+- 6d5e8339: Allow sessionInit overrides and fix useHeadset android chrome bug"
+
 ## 0.0.9
 
 ### Patch Changes